more fixing

+++ src/controllers/products.controller.js
@@ -11,21 +11,21 @@ export const getTopProducts = async (req, res, next) => {
     if (cached) return res.json(cached);
 
     const sql = `
-      SELECT 
-        p.name AS product_name,
-        SUM(ps.quantity) AS total_sold,
-        ROUND(SUM(ps.total_price)::numeric, 2) AS total_revenue,
-        ROUND(AVG(ps.base_price)::numeric, 2) AS avg_price,
-        ROUND(AVG(ps.base_price * $3)::numeric, 2) AS avg_cost,
-        ROUND(((AVG(ps.base_price) - AVG(ps.base_price * $3)) / NULLIF(AVG(ps.base_price), 0)) * 100, 2) AS margin_percent
-      FROM product_sales ps
-      JOIN products p ON p.id = ps.product_id
-      JOIN sales s ON s.id = ps.sale_id
-      WHERE s.created_at BETWEEN $1 AND $2
-      GROUP BY p.name
-      ORDER BY total_revenue DESC
-      LIMIT $4;
-    `;
+  SELECT 
+    p.name AS product_name,
+    SUM(ps.quantity) AS total_sold,
+    ROUND(SUM(ps.total_price)::numeric, 2) AS total_revenue,
+    ROUND(AVG(ps.base_price)::numeric, 2) AS avg_price,
+    ROUND(AVG(ps.base_price * $3)::numeric, 2) AS avg_cost,
+    ROUND((((AVG(ps.base_price) - AVG(ps.base_price * $3)) / NULLIF(AVG(ps.base_price), 0)) * 100)::numeric, 2) AS margin_percent
+  FROM product_sales ps
+  JOIN products p ON p.id = ps.product_id
+  JOIN sales s ON s.id = ps.sale_id
+  WHERE s.created_at BETWEEN $1 AND $2
+  GROUP BY p.name
+  ORDER BY total_revenue DESC
+  LIMIT $4;
+`;
 
     const result = await query(sql, [start, end, cost_pct, limit]);
